Smaller compiler warnings fixes.

+++ src/components/ogre/widgets/ActionBarIconManager.cpp
@@ -56,7 +56,7 @@ ActionBarIconSlot* ActionBarIconManager::createSlot(unsigned int pixelSize) {
 	ss << "actionBarIconSlot" << mSlotsCounter++;
 	//Make the slot more visible.
 	UniqueWindowPtr<CEGUI::Window> container(mGuiManager.createWindow("EmberLook/StaticImage", ss.str()));
-	container->setSize(CEGUI::USize(CEGUI::UDim(0.f, pixelSize), CEGUI::UDim(0.f, pixelSize)));
+	container->setSize(CEGUI::USize(CEGUI::UDim(0.f, (float)pixelSize), CEGUI::UDim(0.f, (float)pixelSize)));
 	auto slot = std::make_unique<ActionBarIconSlot>(std::move(container));
 	mSlots.emplace_back(std::move(slot));
 	return mSlots.back().get();
@@ -76,7 +76,7 @@ ActionBarIcon* ActionBarIconManager::createIcon(Gui::Icons::Icon* icon, unsigned
 	UniqueWindowPtr<CEGUI::Window> iconWindow(mGuiManager.createWindow("EmberLook/StaticImage", ss.str()));
 
 	if (item && iconWindow) {
-		item->setSize(CEGUI::USize(CEGUI::UDim(0.f, pixelSize), CEGUI::UDim(0.f, pixelSize)));
+		item->setSize(CEGUI::USize(CEGUI::UDim(0.f, (float)pixelSize), CEGUI::UDim(0.f, (float)pixelSize)));
 		//item->setTooltipText(name);
 
 		iconWindow->setProperty("BackgroundEnabled", "false");
